Initial support for complete linking and exit handling
Propagate exit events between linked actors. Implement Actor.trap_exit.
Clean up IO objects used by mailboxes/wakers when actors crash.

This is a bit of a megacommit. Mea culpa. So much of this code was interrelated
and dependent on each other I had trouble breaking it apart into smaller
commits. Linking is one of the most complex and harriest parts of the entire
library as there are so many moving parts and it touches so many pieces of
the implementation.

This approach has actors propagate exit events as ExitEvent objects, which are
in turn a subclass of SystemEvent (YAGNI says I can probably get rid of
SystemEvent as it's unlikely there will be other types of SystemEvents).
Exit messages are given a high priority by prepending them on to the front
of a given actor's mailbox. They're handled out of band from normal receive:
when an exit message is encountered, it's raised from receive.

Actor.trap_exit is implemented as a class method which is called when a given
Celluloid::Actor class is defined. This registers a callback that's fired
when exit events are received by instances of that actor. This allows actors
to process exit events and act on them. If actors don't register a trap_exit
handler, they crash when they receive a linked actor's exit event.