Refactor Celluloid::Actor into a class
Celluloid initially began with the idea of augmenting a given class with
actor-like capabilities by including the Celluloid::Actor module. As the
scope of what a Celluloid::Actor had to do grew, this approach was
unteneable. The code soon grew to be littered with underscore prefixed
methods and instance variables. Clearly this is a job for a class!

Later releases of Celluloid moved away from the "include Celluloid::Actor"
API towards the simpler "include Celluloid" API. As of this commit, this is
the ONLY supported API.

This commit changes Celluloid::Actor into a class which wraps a given
subject object. Everything that was previously a wonky underscore method
or underscore instance variable is now a regular method or instance
variable of the Celluloid::Actor class.

Things are still a bit goofy in places after this change, but it represents
the initial work towards cleaner internals. Subsequent commits will focus
on cleaning up any other wonky code which resulted from the previous
way the Celluloid::Actor internals were factored.