Identify calls by ID rather than identical objects
This change is mostly for DCell, which proxies the Celluloid actor protocol over 0MQ.
However, to do this it must marshal/unmarshal objects, so responses contain copies of
the original call object. The previous implementation did a direct object comparison
of call objects, so this behavior was broken in DCell.

This implementation memoizes the object ID and uses that to identify calls. This may
be a bad approach if objects are garbage collected and their IDs reused. More
investigation is needed.