Celluloid#after schedules callbacks which fire after a given interval
Timers are a popular feature of event loops, and work just as well in a
concurrent object system like Celluloid.

This system lets you specify a block which gets fired after a given
interval. Calling Celluloid#after returns a Celluloid::Timer object that
can be cancelled if need be. Otherwise the actor automatically waits the
given amount of time then runs the callback. The callback works with the
same semantics as message sends, running in its own fiber which can be
suspended/resumed as needed.