Extract Celluloid::Fiber into Celluloid::Task
There's a more fundamental abstraction than fibers going on inside of
Celluloid. That abstraction is a task. Every actor contains several
tasks, only one of which can (presently) run at once.

However, this is behavior that might be interesting to experiment with.
Gems like Cloby and the clojure-gem provide some powerful tools for
managing concurrent state, such as STM and immutable persistent data
structures. With these at our disposal it would be possible for all
methods to execute concurrently and still provide localized state and
safe concurrency.

Tasks are also something we should be able to introspect on. In the
commits to follow I'll be adding some interfaces that let you see what
an actor is doing at a given point in time.