Enable introspection on Celluloid::Tasks via Celluloid#tasks
This method introspects on running tasks within an actor, returning a
hash of all tasks to what they're presently doing. So far it only works
for calls since they're the only thing using the new tasks API, but it
should be easy to expand to everything that suspends and resumes tasks.