Have Response objects reference their original calls
Responses were referencing calls by their object IDs in an odd bit of
coupling between Celluloid and DCell. This shouldn't be! If responses
reference their calls, we can tack the fiber associated with the call
onto the call, and resume it directly rather than using signals to track
calls.

Note: this is guaranteed to break DCell in its current form. DCell will
need to be updated.