Store and return an array of tasks with a status attribute
The previous implementation of Celluloid#tasks returned a hash of tasks
to what they were waiting on. The new implementation returns an array of
tasks, each of which has a #status attribute which tells you what the
current task is doing.

This breaks compatibility with both DCell and Celluloid::IO, but no
worries, they'll get updated soon!