Thread-free futures
The past implementation of futures was a bit of a legacy one. The
initial use case was to run a block and obtain the value later, and when
it came time to let you make future calls to actors, this same approach
was used even though a thread is unnecessary in this case as it can all
be handled with mailboxes.

The new approach makes futures work more like mailboxes and participate
directly in the actor protocol. Futures will automatically store a value
when it is received, and can forward that value along to interested
actors which are already waiting on a given value.