Rename Celluloid::Application to Celluloid::Group
"Application" is a name I stole from Erlang, and in the continuing story
of borrowing things from Erlang, the names are always confusing. The
name "group" in Erlang has a crazy meaning related to how error messages
get directed to a person controlling a group of actors.

Group is a much more meaningful name for groups of actors, so group it
is!