Reimplementation that composes with core IO types instead of subclassing
The previous (and sadly largely untested) spike of the Celluloid::IO
replacement types didn't come close to actually working properly and
actually only passed the existing tests by blocking the actor, which is
something the present test doesn't attempt to cover, but should.

To actually get it to work, I don't think it's possible to do with
subclassing. Celluloid::IO::TCPServer#accept and #accept_nonblock need
to return a Celluloid::IO::TCPSocket object, however I don't see any
non-hax way to turn a TCPSocket into a Celluloid::IO::TCPSocket if
Celluloid::IO::TCPSocket is a subclass of TCPSocket to begin with.

This new implementation composes with the core IO types and tries to use
Forwardable to make delegating a bit less painless, but several of the
core methods are still undelegated.