Avoid "ThreadError: Attempt to unlock a mutex which is not locked"
The mailbox avoids using #synchronize for speed (block creation
overhead), however getting the semantics right here is a bit tricky. In
certain situations @mutex.lock will throw an error.

In these cases, we should raise the original exception, however if we
attempt to unlock the mutex it will instead raise this ThreadError.

Moving the locking operation out of the ensure block avoids this
problem.