Round-robin behavior for Celluloid::Pool
Previously actors were popped off of and onto a pool stack. This means
that asking for actors from a pool will always return the same actors
over and over unless there is contention around the pool.

This change gives the pool a more round-robin behavior by pulling from
the front of the array and adding returned actors onto the back.