Updated TCPSocket#read to work with nil length
According to the IO classes, a read method should read the
input to the end when a length of nil, or no length is supplied.
Celluloid::IO::TCPSocket could not handle this case and crashed.
This commit adds behaviour that will read the input until no
more input is available. A spec was added to test this as well