'exclusive' class method
The 'exclusive' method can be used to mark certain instance methods as
always running exclusive. This has the side effect (which so happens to
be the main use case) of running code outside of a Ruby Fiber, allowing
that code to consumer more stack space than the 4kB stacks of YARV
fibers would otherwise allow.