make IncidentLogger a plain old ruby object
This may be reverted eventually, but as I'm trying to fix test failures,
it seems that having the logger itself be managed by the system it's
trying to log is going to cause trouble.

The only change to IncidentLogger itself is the addition of a mutex
around the ring buffers. The only functionality removed is the ability
to log messages asynchronously, but there probably isn't much need for
that.

IncidentReporter is still an actor, as it needs to subscribe to
notifications and only actors can do that.