Add Cellulate, a late init Celluloid mixin
This commit adds Cellulate, a mixin which can be used to convert
an existing object into a Celluloid Actor proxy.

Either mix in, or extend your object with Cellulate and assign it
the result of its own "cellulate" call. See cellulate.rb in
examples for code sample.

TODO:
testing
create Cellulate::IO and DCellulate for the corresponding
libraries.