Ignore dead linked actors when propagating crashes
We may have linked to an actor that has died at the same time we've
died. This leads to a potential race between propagating crash messages
to each other. This manifests in the form of a MailboxError when we
attempt to deliver these crash notifications to other actors.

These exceptions can be safely ignored as the other actor is already
dead and thus incapable of handling crash messages itself. Any
non-MailboxError exceptions are still escalated and will continue to
crash actor cleanup.