SSL accept support
Adds an #accept method to SSLSocket, upgrading the already-accepted TCP
socket to an SSL connection. This is the same API as OpenSSL::SSL::SSLSocket

Also adds a convenience Celluloid::IO::SSLServer class, with the same
API as the OpenSSL::SSL::SSLServer class, which wraps a TCPServer and
automatically upgrades accepted sockets to SSL before returning them.