Only log dangerous suspends if $CELLULOID_DEBUG is set
These warnings aren't helpful and are way too annoying for average users.

We should work on fixing the internal sources of these warnings before we expose
them to average users of the framework.

At the very least we need a way of acknowledging them and hiding them in
situations where they're irrelevant.