Don't use null object as logger in tests
It seems that null objects don't handle `warn` properly, sometimes, or something. This was causing lots of stuff to be printed to stderr in celluloid-zmq's tests. It isn't really necessary to use a null object, we can simply make an assertion on the actual logger object. We've defined the logger to log to a special fine in the spec_helper anyway. This also removes the need to reset the logger in an around filter, which is nice.