Simplify dependencies: merge 'essentials' and 'culture' repos
This merges the current contents of the 'culture' and 'essentials' repos back
into the mainline repository.

Rationale:

- Use of a submodule is confusing

The previous implementation incorporated the "culture" repo as a git submodule.
It also used this submodule for the most essential functions of a repo,
meaning that the repo was effectively in a "broken" state by default with
no guidance as to how to get it working.

- Cyclical dependencies on celluloid-essentials

There's really no purpose for celluloid-essentials to be a separate gem:
It's core functionality of Celluloid which the base gem cannot function without.
This really belongs in the mainline.

- Coupling via global variables between gems

There is presently some rather nasty coupling between gems via global variables
which is very difficult to modify.

In general global variables should be avoided, and they absolutely should not
be used between gems.

Concerns:

- Notifier: Systems which depend on Celluloid::Notifier seem to be having
  problems as of this commit
- IncidentReporter: does not seem to be functioning correctly
- Probe: Seems to be having problems due to IncidentReporter